wrapper

+++ khd/kernels/utils/threads.h
@@ -9,7 +9,7 @@ __device__ uint get_num_blocks() { return gridDim.x * gridDim.y * gridDim.z; }
 
 __device__ uint get_block_id() { return gridDim.x * gridDim.y * blockIdx.z + gridDim.x * blockIdx.y + blockIdx.x; }
 
-__device__ int64_t get_local_thread_id() {
+__device__ uint get_local_thread_id() {
     return blockDim.x * blockDim.y * threadIdx.z + blockDim.x * threadIdx.y + threadIdx.x;
 }
